+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "sync-scroll" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [2.0.0]
+
+- Extension not enabled by default (configurable using `"syncScroll.enabled": true`)
+- Default scroll mode to normal (keep persistence)
+- Bump dependencies
+- Cleanup whole repository (lint + old files)
+
 ## [1.3.0]
 
 - Add command to jump to corresponding position in the next panel